feat: Add links and fix some bugs

+++ users/signals.py
@@ -1,5 +1,5 @@
 from django.contrib.auth.models import User
-from django.db.models.signals import post_save, pre_delete
+from django.db.models.signals import post_save, post_delete
 from django.dispatch import receiver
 
 from .models import Profile, Location
@@ -16,6 +16,7 @@ def create_profile_location(sender, instance, created, **kwargs):
     instance.location=profile_location
     instance.save()
 
-@receiver(pre_delete, sender=Profile)
+@receiver(post_delete, sender=Profile)
 def delete_location_profile(sender, instance, *args, **kwargs):
-  instance.location.delete()
+  if instance.location != None:
+    instance.location.delete()
